Importer Status...
Speed. This is now mostly acceptable. The one exception is pass
'BreakAllCsetCycles'. The reason is that the limit computation it does
for backward branches still uses the inefficient file-level dependency
computation. This will be tackled in short order. For the other passes
the file spent is 'CsetsDeps' is recouped by the much faster graph
setup.

Memory. This seems to be mostly acceptable as well, with the
exceptions of 'BreakAllCsetCycles' (again, for reasons see above), and
'InitCsets'. It seems to happen while the pass breaks internal
dependencies, but there is no hard data. I have to measure using a
memory-debug enabled tclsh. I suspect either the actual internal
dependencies, or the pseudo-dependencies. Maybe combined with a bad
choice of data structures. Well, measuring first.