Implementation of a linked list to solve the memory leak described in a TODO in [e2ebb1f5cae8].
This code is slower than having the memory leak, and at the end, it was not a big memory leak. Let's say, 10 byte per revision involved in a file annotate. If a file has 30000 revisions, it may go around 300KB then.

For this leak to be noticeable (as we have a content cache until 50MB), it would mean having above a milion revisions for the file annotated.

So, I keep this in a branch apart, as maybe the leak is worth the faster code. I have not measured the speed difference though.