Adding a more options at naming commits for diff and update commands.
I wrote a new function that allows more kind of specifications for a checkin uuid: checkout,
parent and pivot:id1:id2.

If there was a way to discover the 'parent branch' of the current checkout, there could be a
'parentbranch' naming too.

I think this makes the life easier, specially for the pivot:id1:id2 case, about seeing
the difference between a branch and its parent branch. I find myself using often the
test-find-pivot command just to get that diff or update to the pivot, to check changes
between the current branch and the parent branch.