Should fix ticket [f0f9aff371f2].
Until now, the annotate_cmd was taking the last (or any?) checkin that had the
artifact to be annotated, totally unrelated to what version is checked out.

I made annotate_cmd respect the checkout, and annotate only from the past until
the checked out version. This makes the command slower, but at least does what
I'd like it to do. Maybe someone else can come up with a faster or simpler
implementation.

I changed the compute_direct_ancestors code so I could call it twice in a single
run.