Fix problem when using "fossil cat" when specifying repository via command line argument (with -R) (reported on ML)
Problem was the use of file_tree_name() which call db_must_be_within_tree().

Add a variable in the Global structure 'g' to remember if -R|--repository
argument was specified and don't call file_tree_name() if it's the case (since
user expect file relative to repository.

(Pending review...)