Add --import-marks and --export-marks options to 'fossil import' subcommand.
This allows one to save all git-generated marks following an import.  Previously,
performing an incremental import from git resulted in new commits being imported
as orphans, and trying to perform an export to git at some later time resulted
in git complaining of undefined marks. To remedy this, the format of the marks
file is amended to include the rid, mark, and UUID of each commit. Now, fossil
and git agree on the mapping between commits and marks, so updates in either
direction can be performed easily.